Update Auth system
