+++ client/src/shared/context/useAuth.js
@@ -1,0 +1,6 @@
+import { useContext } from 'react'
+import { AuthContext } from '@/shared/context/AuthContext'
+
+export function useAuth() {
+    return useContext(AuthContext)
+}
